[larger-key-info] Extract weight and sets display into files
