+++ app/src/main/java/com/litus_animae/refitted/compose/exercise/ExerciseSet.kt
@@ -1,8 +1,5 @@
 package com.litus_animae.refitted.compose.exercise
 
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.ColumnScope
 import androidx.compose.foundation.layout.Row
@@ -11,11 +8,8 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material.Button
 import androidx.compose.material.Card
-import androidx.compose.material.Icon
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Text
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.rounded.Edit
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.derivedStateOf
 import androidx.compose.runtime.getValue
@@ -94,68 +88,14 @@ fun ColumnScope.ExerciseSetView(
           .padding(bottom = 5.dp)
           .weight(1f)
       ) {
-        Column(
-          Modifier
-            .fillMaxWidth()
-            .padding(5.dp)
-            .clickable { onStartEditWeight(weight) }) {
-          Icon(
-            Icons.Rounded.Edit,
-            contentDescription = null,
-            Modifier.align(Alignment.End)
-          )
-        }
-        Column(
-          Modifier.fillMaxWidth(),
-          verticalArrangement = Arrangement.Center,
-          horizontalAlignment = Alignment.CenterHorizontally
-        ) {
-          val weightLabel = stringResource(id = R.string.weight_label)
-          val weightUnit = stringResource(id = R.string.lbs)
-          Text(
-            weightLabel,
-            style = MaterialTheme.typography.h4,
-            modifier = Modifier.align(Alignment.CenterHorizontally)
-          )
-          val displayWeight = String.format("%.1f", saveWeight)
-          Text(
-            "$displayWeight $weightUnit",
-            style = MaterialTheme.typography.h5,
-            modifier = Modifier.align(Alignment.CenterHorizontally)
-          )
-        }
+        WeightDisplay(onStartEditWeight, weight, saveWeight)
       }
       Card(
         Modifier
           .fillMaxWidth()
           .weight(1f)
       ) {
-        Box(
-          Modifier.fillMaxSize(),
-        ) {
-          Column(
-            Modifier.fillMaxSize(),
-            verticalArrangement = Arrangement.Center,
-            horizontalAlignment = Alignment.CenterHorizontally
-          ) {
-            val toCompletionSetPhrase = optionWhen(exerciseSet.sets < 0) {
-              val expectedReps = exerciseSet.reps(numCompleted)
-              if (expectedReps < 0) {
-                // This is a "complete as many as possible" set
-                "${record.cumulativeReps}"
-              } else {
-                // This is a "complete this many reps in as many sets as you need" set
-                "${expectedReps - record.cumulativeReps} / $expectedReps"
-              } to "Reps Completed"
-            }
-            val (setDisplay, setSubtext) = toCompletionSetPhrase
-              .getOrElse {
-                "$numCompleted / ${exerciseSet.sets}" to "Sets Completed"
-              }
-            Text(setDisplay, style = MaterialTheme.typography.h4)
-            Text(setSubtext, style = MaterialTheme.typography.h5)
-          }
-        }
+        SetsDisplay(exerciseSet, numCompleted, record)
       }
     }
     Column(Modifier.weight(1f), horizontalAlignment = Alignment.End) {
